@@ -16,7 +16,6 @@ use Pluswerk\CacheAutomation\Service\Configuration;
 use TYPO3\CMS\Core\DataHandling\DataHandler;
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Object\ObjectManager;
 use TYPO3\CMS\Extbase\Service\CacheService;
 use RuntimeException;
 
@@ -26,46 +25,36 @@ use RuntimeException;
  * @author Markus Hölzle <markus.hoelzle@pluswerk.ag>
  * @package Pluswerk\CacheAutomation\Hook
  */
-class DataHandlerDetector implements SingletonInterface
+final readonly class DataHandlerDetector implements SingletonInterface
 {
-
-    /**
-     * @var Configuration
-     */
-    protected $configuration;
-
-    /**
-     * @var CacheService
-     */
-    protected $cacheService;
+    private Configuration $configuration;
 
     /**
      * DataHandlerHook constructor.
      */
-    public function __construct()
+    public function __construct(private CacheService $cacheService)
     {
         $this->configuration = Configuration::getInstance();
-        $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-        $this->cacheService = $objectManager->get(CacheService::class);
     }
 
     /**
      * @param string $status The status of the record
      * @param string $table Database table name
      * @param int|string $id The uid of the record or something like "NEW59785a1ec52" if the record is new
-     * @param array $changedFields Field value map of the changed fields
+     * @param array<mixed> $changedFields Field value map of the changed fields
      * @param DataHandler $dataHandler Reference back to the DataHandler
      * @throws \RuntimeException
      */
     // @phpcs:ignore PSR1.Methods.CamelCapsMethodName
-    public function processDatamap_afterDatabaseOperations(/** @noinspection PhpUnusedParameterInspection */ string $status, string $table, $id, array $changedFields, DataHandler $dataHandler): void
+    public function processDatamap_afterDatabaseOperations(/** @noinspection PhpUnusedParameterInspection */ string $status, string $table, int|string $id, array $changedFields, DataHandler $dataHandler): void
     {
         $expiredPages = [];
         if ($this->configuration->isConfigured($table)) {
             $agentConfigurations = $this->configuration->getAgentsForTable($table);
             foreach ($agentConfigurations as $agentConfiguration) {
-                /** @var AgentInterface $agent */
-                $agent = GeneralUtility::makeInstance($agentConfiguration['agent']);
+                $agentClass = $agentConfiguration['agent'];
+                /** @var class-string $agentClass */
+                $agent = GeneralUtility::makeInstance($agentClass);
                 if ($agent instanceof AgentInterface) {
                     $expiredPages = $agent->getExpiredPages($table, $id, $agentConfiguration['agentConfiguration'], $changedFields);
                 } else {
@@ -73,6 +62,7 @@ class DataHandlerDetector implements SingletonInterface
                 }
             }
         }
+
         if (count($expiredPages) !== 0) {
             $this->cacheService->clearPageCache(array_unique($expiredPages));
         }
